!fix: add chainID validation to keyAuth (#1078)
This is needed to add replay protection mainly between tempo testnets
and mainnets, and maybe other chains that adopt this in the future.

Note: we still don't have nonce replay protection mentioned in #1033 , I
think the cleaner way to do this, is to just not allow re-authorization
of keys that have been revoked.

This will come in a follow up PR, which will make some changes to how
active keys work.