fix!(precompiles): add explicit invalidToken() error  (#893)
Currently, the TIP20::from_address function, does not check if the
address actually has the TIP20 prefix.
This leads to random errors like PrecompileOOG, when it is used in
places like the StablecoinDex.
Eg. https://github.com/tempoxyz/docs/pull/182#discussion_r2515563659

We add an InvalidToken() error specifically for this case in this PR.

TBD: @0xKitsune is there a reason why we use unchecked address to
tokenId conversion here -
https://github.com/tempoxyz/tempo/blob/main/crates/precompiles/src/tip20/mod.rs#L640

If not, then we could move this invalid token check inside the TIP20 fns